ba_rules/requirements.txt
durable-rules==2.0.28
ipython==7.17.0